+++ apps/app/src/components/ScheduleCard/index.ts
